pyproject.toml
[tool.poetry]
name = "mittari-server"
version = "0.1.0"
description = ""
authors = ["Markus Juopperi"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
pydantic = "^1.9.1"
redis = "^4.3.4"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
python-dotenv = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app:main"
